[Clickhouse Writer] Handle big transactions correctly
When we have large amounts of data and we go over max_query_size, we
can't use batched inserts with values lists, because that will cause
incorrect data to be inserted if we're using aggregations with GROUP BY.
Rows that should be grouped together can end up in different batches,
messing things up.

We're now inserting data that comes from input.sql to a temporary table
and then running output.sql on that once all of the data is written out
making it a single query, which fixes the issue with grouping and aggs.